Port espBode to ESP32-C3 using W5500 Ethernet and explicit AWG UART wiring.
Replace WiFi/ESPTelnet dependencies with Ethernet-compatible network wrappers, add a native telnet server, and route AWG traffic through Serial1 with configurable pins for reliable pass-through and readback.

Made-with: Cursor

+++ vxi_server.cpp
@@ -51,7 +51,7 @@ void VXI_Server::begin ( bool bNext )
     vxi_port++;
   }
   
-  tcp_server.begin(vxi_port);
+  tcp_server.begin((uint16_t)vxi_port);
 
   Debug.Progress() << "\nListening for VXI commands on TCP port " << vxi_port << "\n";
 }
